[difftest] spike support
