+++ difftest/Cargo.toml
@@ -1,7 +1,22 @@
 [workspace]
 resolver = "2"
 members = [
-  "libspike_rs",
-  "offline",
-  "online_drive",
+    "test_common",
+    "libspike_rs",
+    "offline",
+    "online_drive",
 ]
+exclude = [
+    "libspike_interfaces"
+]
+
+[workspace.package]
+version = "0.1.0"
+
+[workspace.dependencies]
+anyhow = "1.0.79"
+clap = { version = "4.4.18", features = ["derive"] }
+tracing = "0.1.40"
+tracing-subscriber = { version = "0.3", features = ["env-filter", "ansi"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
