+++ difftest/libspike_rs/Cargo.toml
@@ -4,8 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+anyhow = { workspace = true }
+tracing = { workspace = true }
 libc = "0.2.155"
-anyhow = "1.0.79"
-tracing = "0.1.40"
-lazy_static = "1.5.0"
 xmas-elf = "0.9.1"
