+++ difftest/offline/Cargo.toml
@@ -4,14 +4,15 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-clap = { version = "4.4.18", features = ["derive"] }
+clap = { workspace = true }
+tracing = { workspace = true }
+tracing-subscriber = { workspace = true }
+anyhow = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+
 libloading = "0.8.1"
 xmas-elf = "0.9.1"
-tracing = "0.1.40"
-tracing-subscriber = { version = "0.3", features = ["env-filter"] }
-anyhow = "1.0.79"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-lazy_static = "1.4.0"
 
+common = { path = "../test_common" }
 libspike_rs = { path = "../libspike_rs" }
